restructured the file structure and added some files in the -util folder
